+++ Kernel/Modules/AdminSLA.pm
@@ -326,6 +326,9 @@ sub _MaskNew {
         $SLAData{ServiceID} = $Self->{ParamObject}->GetParam( Param => 'ServiceID' );
     }
 
+    # get list type
+    my $ListType = $Self->{ConfigObject}->Get('Ticket::Frontend::ListType');
+
     # get service list
     my %ServiceList = $Self->{ServiceObject}->ServiceList(
         Valid  => 1,
@@ -340,6 +343,7 @@ sub _MaskNew {
         Multiple    => 1,
         Size        => 5,
         Translation => 0,
+        TreeView    => ($ListType eq 'tree') ? 1 : 0,
         Max         => 200,
     );