+++ Kernel/Modules/AgentStats.pm
@@ -1594,6 +1594,7 @@ sub Run {
                     Name           => $ObjectAttribute->{Element},
                     Multiple       => 1,
                     Size           => 5,
+                    Class          => ($ObjectAttribute->{ShowAsTree} && $ObjectAttribute->{IsDynamicField}) ? 'DynamicFieldWithTreeView' : '',
                     SelectedID     => $ObjectAttribute->{SelectedValues},
                     Translation    => $ObjectAttribute->{Translation},
                     TreeView       => $ObjectAttribute->{TreeView} || 0,
@@ -1602,6 +1603,10 @@ sub Run {
                     OnChange       => "Core.Agent.Stats.SelectCheckbox('Select"
                         . $ObjectAttribute->{Element} . "')",
                 );
+
+                if ($ObjectAttribute->{ShowAsTree} && $ObjectAttribute->{IsDynamicField}) {
+                    $BlockData{SelectField} .= ' <a href="#" title="$Text{"Show Tree Selection"}" class="ShowTreeSelection">$Text{"Show Tree Selection"}</a>';
+                }
             }
 
             $BlockData{Name}    = $ObjectAttribute->{Name};
@@ -1717,6 +1722,7 @@ sub Run {
                     Name           => $ObjectAttribute->{Element},
                     Multiple       => 1,
                     Size           => 5,
+                    Class          => ($ObjectAttribute->{ShowAsTree} && $ObjectAttribute->{IsDynamicField}) ? 'DynamicFieldWithTreeView' : '',
                     SelectedID     => $ObjectAttribute->{SelectedValues},
                     Translation    => $ObjectAttribute->{Translation},
                     TreeView       => $ObjectAttribute->{TreeView} || 0,
@@ -1725,6 +1731,10 @@ sub Run {
                     OnChange       => "Core.Agent.Stats.SelectCheckbox('Select"
                         . $ObjectAttribute->{Element} . "')",
                 );
+
+                if ($ObjectAttribute->{ShowAsTree} && $ObjectAttribute->{IsDynamicField}) {
+                    $BlockData{SelectField} .= ' <a href="#" title="$Text{"Show Tree Selection"}" class="ShowTreeSelection">$Text{"Show Tree Selection"}</a>';
+                }
             }
 
             $BlockData{Element} = $ObjectAttribute->{Element};